# docs/site/public/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "description": "FastAPI's error envelope \u2014 `{\"detail\": \"...\"}` \u2014 named so the schema can\npoint at it and a generated client can read it.",
        "properties": {
          "detail": {
            "description": "a sentence for a developer, never for an end user",
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlatformCommandsResponse": {
        "description": "What the whole deployment's users actually type.",
        "properties": {
          "commands": {
            "items": {
              "$ref": "#/components/schemas/PlatformCommandRow"
            },
            "title": "Commands",
            "type": "array"
          },
          "end": {
            "format": "date",
            "title": "End",
            "type": "string"
          },
          "start": {
            "format": "date",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end",
          "commands"
        ],
        "title": "PlatformCommandsResponse",
        "type": "object"
      },
      "PlatformDailyResponse": {
        "description": "The deployment's activity, one row per day.",
        "properties": {
          "days": {
            "items": {
              "$ref": "#/components/schemas/PlatformDayRow"
            },
            "title": "Days",
            "type": "array"
          },
          "end": {
            "format": "date",
            "title": "End",
            "type": "string"
          },
          "start": {
            "format": "date",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end",
          "days"
        ],
        "title": "PlatformDailyResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "The HTTP surface behind the Telegram Mini App and the web console.",
    "title": "Cookiebot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/analytics/commands": {
      "get": {
        "description": "Which commands the deployment uses, busiest first, with how many groups\neach one reaches \u2014 a command with 10,000 invocations in one group and one\nused everywhere are different facts, and a decision to retire a command\nshould not confuse them.",
        "operationId": "commands_admin_analytics_commands_get",
        "parameters": [
          {
            "description": "inclusive, UTC",
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "inclusive, UTC",
              "title": "Start"
            }
          },
          {
            "description": "inclusive, UTC",
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "inclusive, UTC",
              "title": "End"
            }
          },
          {
            "description": "how many commands to return",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "how many commands to return",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformCommandsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "the window is reversed or longer than a year"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "no bearer token, or one that did not verify"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "not an owner of this deployment \u2014 or an owner whose token lacks the scope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Which commands the whole deployment uses",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/analytics/daily": {
      "get": {
        "description": "One row per day, summed across every group that was active that day.\n\n`active_users` is summed per group, so one person in three groups counts\nthree times \u2014 deduplicating across groups would need the raw events rather\nthan the rollups, which is a much more expensive question.",
        "operationId": "daily_admin_analytics_daily_get",
        "parameters": [
          {
            "description": "inclusive, UTC",
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "inclusive, UTC",
              "title": "Start"
            }
          },
          {
            "description": "inclusive, UTC",
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "inclusive, UTC",
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformDailyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "the window is reversed or longer than a year"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "no bearer token, or one that did not verify"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "not an owner of this deployment \u2014 or an owner whose token lacks the scope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "One row per day, summed across every group",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/analytics/groups": {
      "get": {
        "description": "The busiest groups in the window, most messages first.\n\nA leaderboard, not a directory: a group with no activity in the window is\nabsent here and present in `/admin/groups`.",
        "operationId": "top_groups_admin_analytics_groups_get",
        "parameters": [
          {
            "description": "inclusive, UTC",
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "inclusive, UTC",
              "title": "Start"
            }
          },
          {
            "description": "inclusive, UTC",
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "inclusive, UTC",
              "title": "End"
            }
          },
          {
            "description": "how many groups to return",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "how many groups to return",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopGroupsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "the window is reversed or longer than a year"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "no bearer token, or one that did not verify"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "not an owner of this deployment \u2014 or an owner whose token lacks the scope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "The busiest groups in the window",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/analytics/llm": {
      "get": {
        "description": "Fleet-wide LLM spend, per provider and model, most expensive first.",
        "operationId": "llm_admin_analytics_llm_get",
        "parameters": [
          {
            "description": "inclusive, UTC",
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "inclusive, UTC",
              "title": "Start"
            }
          },
          {
            "description": "inclusive, UTC",
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "inclusive, UTC",
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformLlmResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "the window is reversed or longer than a year"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "no bearer token, or one that did not verify"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "not an owner of this deployment \u2014 or an owner whose token lacks the scope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "What every group's AI features cost together",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/admin/tenant": {
      "get": {
        "description": "The tenant's own row, read-only.\n\nRead-only because every field here changes how the *bot* behaves \u2014 which\ncommands exist, which model answers, where media is written \u2014 and a Mini\nApp form that could flip `active` or empty `owner_ids` would be one\nmis-tap from an outage nobody could undo through the same API. Editing a\ntenant is a deliberate database change, and it stays one.\n\n`bot_tokens` is deliberately absent: an owner who needs it has it already,\nand an endpoint that returns bot tokens is one stolen owner token away\nfrom being the whole deployment.",
        "operationId": "tenant_admin_tenant_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "no bearer token, or one that did not verify"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "not an owner of this deployment \u2014 or an owner whose token lacks the scope"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "How this deployment is configured",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
...
}